refactor: header folder name, storybook fixed
