+++ src/constants/stories.constants.ts
@@ -28,7 +28,7 @@ import { IMobileSearchbar } from '../components/UI/molecules/MobileSearchbar'
 import { IListMenuItems, MenuItemInfo } from '../components/UI/molecules/ListMenuItems/ListMenuItems.interface'
 import { IMenuItems } from '../components/UI/molecules/ListMenuIcons/ListMenuIcons.interface'
 import { IHeaderDrawerTabs } from '../components/UI/molecules/HeaderDrawerTabs/HeaderDrawerTabs.interface'
-import { ILogoutHeader } from '../components/UI/organism/LogOutHeader'
+import { ILogoutHeader } from '../components/UI/organism/LogoutHeader'
 import { IAvatar } from '../components/UI/atoms/Avatar'
 import { IHeaderTabs } from '../components/UI/molecules/HeaderTabs/HeaderTabs.interface'
 import { IHeaderTab } from '../components/UI/atoms/HeaderTab'
