+++ src/index.ts
@@ -1,2 +1,2 @@
 export * from './components/UI/template/logoutTemplate'
-export * from './components/UI/organism/LogOutHeader'
+export * from './components/UI/organism/LogoutHeader'
